SIM: Fixed bug that had no effect (remove \n, not spaces).
Inside GameMap.arrayToText(), we use rstrip() after strOut has already
been contructed by a loop. We use it to remove the last trailing new-line
charater, in order to use split("\n") on the strOut and get the accurate
number of lines.

Truned out, we forgot to pass "\n" to rstrip(), which lead rstrip() to still
do its job (and remove \n, as well as trailing spaces - of which there weren't
any).

This "bug" lead us the another actual bug in Simulator.draw() that is somewhat
related to the padding feature of arrayToText().