SIM: Fix malfunctioning sim.draw() where player map not properly displayed.
The problem was that the player map (on the right side), actually the last
line of the map, was moved towards the left into the sim map. This was because
we used rstrip() on the text of the sim map (with the intention of removing the
last new-line character. But, we forgot to pass "\n" to rstrip, which in turn
removed all the spaces that have been added by GameMap.toText(pad=True) before.
Hence, padding was working but destroyed by Sim.draw().

All this was fixed by adding "\n" as argument to rstrip() when called on the sim
map text.

In the process we introduced GameMap.toText() (to wrap arrayToText()).

Fixes #12